Refactoring, migration to fs lib and add local storage protocol
