+++ external_file_location/tasks/__init__.py
@@ -20,5 +20,7 @@
 #
 ###############################################################################
 
+from . import abstract_fs
 from . import ftp
 from . import sftp
+from . import filestore
